ai-enhanced-qa-backend/scripts/requirements.txt
gensim
nltk
pinecone
python-dotenv